Move console args into ApplicationArguments.cs instead of IUI classes
