+++ TagsCloudContainer/UI/IUI.cs
@@ -2,6 +2,5 @@ namespace TagsCloudContainer.UI;
 
 public interface IUI
 {
-    ApplicationArguments Setup(string[] args);
     void View(string output);
 }
